backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
PyPDF2==3.0.1
SQLAlchemy==2.0.35
Werkzeug==3.0.4
python-dotenv==1.0.1
requests==2.32.3
openai==0.28.0
python-docx==1.1.2
pytesseract==0.3.13
pytest==8.3.3
coverage==7.6.1
pypdf==5.0.1